python_utilities/requirements.txt
google-cloud-firestore>=2.11.0
google-auth>=2.0.0
requests>=2.28.0
